requirements.txt
# F1 Data Sources
fastf1>=3.6.0
requests>=2.31.0

# ML & Data Science
xgboost>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Database
supabase>=2.0.0
python-dotenv>=1.0.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
